Move fieldset to _core
