+++ tests/test_diffusion.py
@@ -6,7 +6,7 @@ from scipy import stats
 
 from parcels._datasets.structured.generated import simple_UV_dataset
 from parcels import Field, VectorField
-from parcels.fieldset import FieldSet
+from parcels import FieldSet
 from parcels.interpolators import XLinear
 from parcels.kernels import AdvectionDiffusionEM, AdvectionDiffusionM1, DiffusionUniformKh
 from parcels.particle import Particle, Variable
