+++ tests/test_interpolation.py
@@ -6,7 +6,7 @@ from parcels._datasets.structured.generated import simple_UV_dataset
 from parcels._datasets.unstructured.generic import datasets as datasets_unstructured
 from parcels._index_search import _search_time_index
 from parcels import Field, VectorField
-from parcels.fieldset import FieldSet
+from parcels import FieldSet
 from parcels.interpolators import (
     UXPiecewiseLinearNode,
     XFreeslip,
